Created Selenium Navigation to HackerRank Submissions
Removed the necessity to have a text file containing login information since that is not very secure for user.
Simply asking user to input username + password. (may change in the future)
Developed the Selenium automation to navigate to the submissions page
Grabbed all the links to each submitted code to be scraped from